feat(feature-manager): remove sms from feature manager (#2218)
* feat(feature-manager): remove sms from feature manager

* build: add mock defaults for Twilio in dev environment
