+++ docker-compose.yml
@@ -55,10 +55,11 @@ services:
       # Keep in sync with the development key in
       # https://github.com/opengovsg/formsg-javascript-sdk/blob/develop/src/resource/signing-keys.ts
       - SIGNING_SECRET_KEY=HDBXpu+2/gu10bLHpy8HjpN89xbA6boH9GwibPGJA8BOXmB+zOUpxCP33/S5p8vBWlPokC7gLR0ca8urVwfMUQ==
-      - TWILIO_ACCOUNT_SID
-      - TWILIO_API_KEY
-      - TWILIO_API_SECRET
-      - TWILIO_MESSAGING_SERVICE_SID
+      # Mock Twilio credentials. SMSes do not work in dev environment.
+      - TWILIO_ACCOUNT_SID=ACmockTwilioAccountSid
+      - TWILIO_API_KEY=mockTwilioApiKey
+      - TWILIO_API_SECRET=mockTwilioApiSecret
+      - TWILIO_MESSAGING_SERVICE_SID=mockTwilioMsgSrvcSid
       - SES_PASS
       - SES_USER
       - OTP_LIFE_SPAN
