+++ src/app/config/feature-manager/index.ts
@@ -1,5 +1,4 @@
 import FeatureManager from './util/FeatureManager.class'
-import sms from './sms.config'
 import spcpMyInfo from './spcp-myinfo.config'
 
 export * from './types'
@@ -8,6 +7,5 @@ const featureManager = new FeatureManager()
 
 // Register features and associated middleware/fallbacks
 featureManager.register(spcpMyInfo)
-featureManager.register(sms)
 
 export default featureManager
