+++ src/app/services/sms/__tests__/sms.factory.spec.ts
@@ -3,12 +3,7 @@ import { okAsync } from 'neverthrow'
 import { mocked } from 'ts-jest/utils'
 import Twilio from 'twilio'
 
-import {
-  FeatureNames,
-  ISms,
-  RegisteredFeature,
-} from 'src/app/config/feature-manager'
-import { MissingFeatureError } from 'src/app/modules/core/core.errors'
+import { ISms } from 'src/app/config/feature-manager/sms.config'
 
 import { createSmsFactory } from '../sms.factory'
 import * as SmsService from '../sms.service'
@@ -40,147 +35,79 @@ const MOCK_BOUNCE_SMS_PARAMS: BounceNotificationSmsParams = {
 describe('sms.factory', () => {
   beforeEach(() => jest.clearAllMocks())
 
-  describe('sms feature disabled', () => {
-    const MOCK_DISABLED_SMS_FEATURE: RegisteredFeature<FeatureNames.Sms> = {
-      isEnabled: false,
-      props: {} as ISms,
-    }
+  const MOCK_SMS_FEATURE: ISms = {
+    twilioAccountSid: 'ACrandomTwilioSid',
+    twilioApiKey: 'SKrandomTwilioAPIKEY',
+    twilioApiSecret: 'this is a super secret',
+    twilioMsgSrvcSid: 'formsg-is-great-pleasehelpme',
+  }
+  const expectedTwilioConfig: TwilioConfig = {
+    msgSrvcSid: MOCK_SMS_FEATURE.twilioMsgSrvcSid,
+    client: MOCKED_TWILIO,
+  }
+  const SmsFactory = createSmsFactory(MOCK_SMS_FEATURE)
 
-    const SmsFactory = createSmsFactory(MOCK_DISABLED_SMS_FEATURE)
+  it('should call SmsService counterpart when invoking sendAdminContactOtp', async () => {
+    // Arrange
+    MockSmsService.sendAdminContactOtp.mockReturnValueOnce(okAsync(true))
 
-    it('should return MissingFeatureError when invoking sendAdminContactOtp', async () => {
-      // Act
-      const result = await SmsFactory.sendAdminContactOtp(
-        'anything',
-        'anything',
-        'anything',
-      )
+    const mockArguments: Parameters<typeof SmsFactory.sendAdminContactOtp> = [
+      'mockRecipient',
+      'mockOtp',
+      'mockFormId',
+    ]
 
-      // Assert
-      expect(result._unsafeUnwrapErr()).toEqual(
-        new MissingFeatureError(FeatureNames.Sms),
-      )
-    })
+    // Act
+    await SmsFactory.sendAdminContactOtp(...mockArguments)
 
-    it('should return MissingFeatureError when invoking sendVerificationOtp', async () => {
-      // Act
-      const result = await SmsFactory.sendVerificationOtp(
-        'anything',
-        'anything',
-        'anything',
-      )
-
-      // Assert
-      expect(result._unsafeUnwrapErr()).toEqual(
-        new MissingFeatureError(FeatureNames.Sms),
-      )
-    })
-
-    it('should return MissingFeatureError when invoking sendFormDeactivatedSms', async () => {
-      // Act
-      const result = await SmsFactory.sendFormDeactivatedSms(
-        MOCK_BOUNCE_SMS_PARAMS,
-      )
-
-      // Assert
-      expect(result._unsafeUnwrapErr()).toEqual(
-        new MissingFeatureError(FeatureNames.Sms),
-      )
-    })
-
-    it('should return MissingFeatureError when invoking sendBouncedSubmissionSms', async () => {
-      // Act
-      const result = await SmsFactory.sendBouncedSubmissionSms(
-        MOCK_BOUNCE_SMS_PARAMS,
-      )
-
-      // Assert
-      expect(result._unsafeUnwrapErr()).toEqual(
-        new MissingFeatureError(FeatureNames.Sms),
-      )
-    })
+    // Assert
+    expect(MockSmsService.sendAdminContactOtp).toHaveBeenCalledTimes(1)
+    expect(MockSmsService.sendAdminContactOtp).toHaveBeenCalledWith(
+      ...mockArguments,
+      expectedTwilioConfig,
+    )
   })
 
-  describe('sms feature enabled', () => {
-    const MOCK_ENABLED_SMS_FEATURE: Required<
-      RegisteredFeature<FeatureNames.Sms>
-    > = {
-      isEnabled: true,
-      props: {
-        twilioAccountSid: 'ACrandomTwilioSid',
-        twilioApiKey: 'SKrandomTwilioAPIKEY',
-        twilioApiSecret: 'this is a super secret',
-        twilioMsgSrvcSid: 'formsg-is-great-pleasehelpme',
-      },
-    }
-    const expectedTwilioConfig: TwilioConfig = {
-      msgSrvcSid: MOCK_ENABLED_SMS_FEATURE.props.twilioMsgSrvcSid,
-      client: MOCKED_TWILIO,
-    }
-    const SmsFactory = createSmsFactory(MOCK_ENABLED_SMS_FEATURE)
+  it('should call SmsService counterpart when invoking sendVerificationOtp', async () => {
+    // Arrange
+    MockSmsService.sendVerificationOtp.mockReturnValue(okAsync(true))
 
-    it('should call SmsService counterpart when invoking sendAdminContactOtp', async () => {
-      // Arrange
-      MockSmsService.sendAdminContactOtp.mockReturnValueOnce(okAsync(true))
+    const mockArguments: Parameters<typeof SmsFactory.sendAdminContactOtp> = [
+      'mockRecipient',
+      'mockOtp',
+      'mockUserId',
+    ]
 
-      const mockArguments: Parameters<typeof SmsFactory.sendAdminContactOtp> = [
-        'mockRecipient',
-        'mockOtp',
-        'mockFormId',
-      ]
+    // Act
+    await SmsFactory.sendVerificationOtp(...mockArguments)
 
-      // Act
-      await SmsFactory.sendAdminContactOtp(...mockArguments)
+    // Assert
+    expect(MockSmsService.sendVerificationOtp).toHaveBeenCalledTimes(1)
+    expect(MockSmsService.sendVerificationOtp).toHaveBeenCalledWith(
+      ...mockArguments,
+      expectedTwilioConfig,
+    )
+  })
 
-      // Assert
-      expect(MockSmsService.sendAdminContactOtp).toHaveBeenCalledTimes(1)
-      expect(MockSmsService.sendAdminContactOtp).toHaveBeenCalledWith(
-        ...mockArguments,
-        expectedTwilioConfig,
-      )
-    })
+  it('should call SmsService when sendFormDeactivatedSms is called', async () => {
+    MockSmsService.sendFormDeactivatedSms.mockReturnValue(okAsync(true))
 
-    it('should call SmsService counterpart when invoking sendVerificationOtp', async () => {
-      // Arrange
-      MockSmsService.sendVerificationOtp.mockResolvedValue(true)
+    await SmsFactory.sendFormDeactivatedSms(MOCK_BOUNCE_SMS_PARAMS)
 
-      const mockArguments: Parameters<typeof SmsFactory.sendAdminContactOtp> = [
-        'mockRecipient',
-        'mockOtp',
-        'mockUserId',
-      ]
+    expect(MockSmsService.sendFormDeactivatedSms).toHaveBeenCalledWith(
+      MOCK_BOUNCE_SMS_PARAMS,
+      expectedTwilioConfig,
+    )
+  })
 
-      // Act
-      await SmsFactory.sendVerificationOtp(...mockArguments)
+  it('should call SmsService when sendBouncedSubmissionSms is called', async () => {
+    MockSmsService.sendBouncedSubmissionSms.mockReturnValue(okAsync(true))
 
-      // Assert
-      expect(MockSmsService.sendVerificationOtp).toHaveBeenCalledTimes(1)
-      expect(MockSmsService.sendVerificationOtp).toHaveBeenCalledWith(
-        ...mockArguments,
-        expectedTwilioConfig,
-      )
-    })
+    await SmsFactory.sendBouncedSubmissionSms(MOCK_BOUNCE_SMS_PARAMS)
 
-    it('should call SmsService when sendFormDeactivatedSms is called', async () => {
-      MockSmsService.sendFormDeactivatedSms.mockResolvedValueOnce(true)
-
-      await SmsFactory.sendFormDeactivatedSms(MOCK_BOUNCE_SMS_PARAMS)
-
-      expect(MockSmsService.sendFormDeactivatedSms).toHaveBeenCalledWith(
-        MOCK_BOUNCE_SMS_PARAMS,
-        expectedTwilioConfig,
-      )
-    })
-
-    it('should call SmsService when sendBouncedSubmissionSms is called', async () => {
-      MockSmsService.sendBouncedSubmissionSms.mockResolvedValueOnce(true)
-
-      await SmsFactory.sendBouncedSubmissionSms(MOCK_BOUNCE_SMS_PARAMS)
-
-      expect(MockSmsService.sendBouncedSubmissionSms).toHaveBeenCalledWith(
-        MOCK_BOUNCE_SMS_PARAMS,
-        expectedTwilioConfig,
-      )
-    })
+    expect(MockSmsService.sendBouncedSubmissionSms).toHaveBeenCalledWith(
+      MOCK_BOUNCE_SMS_PARAMS,
+      expectedTwilioConfig,
+    )
   })
 })
